Added tabs to sidesheet
