+++ sidesheet/src/components/types.tsx
@@ -1,3 +1,4 @@
 export interface Tab {
   TabTitle: string;
+  TabContent: JSX.Element;
 }
