+++ sidesheet/src/index.tsx
@@ -1,1 +1,2 @@
 export { ProcosysSideSheet } from "./components/sidesheet";
+export type { Tab } from "./components/types";
